% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_Cell_Types.R
\name{annotateCellTypes}
\alias{annotateCellTypes}
\title{Cell Type Annotation with SingleR \link{CCBR} \link{scRNA-seq}}
\usage{
annotateCellTypes(
  object,
  species = "Mouse",
  reduction.type = "umap",
  legend.dot.size = 2,
  do.finetuning = FALSE,
  local.celldex = NULL,
  use.clusters = NULL
)
}
\arguments{
\item{object}{Object of class Seurat (your combined Seurat Object
after PC reduction has been performed)}

\item{species}{The species of your samples ("Human" or "Mouse").
Default is "Mouse"}

\item{reduction.type}{Select the kind of clustering visualization you would
like to use to visualize the cell type results ("umap", "tsne", "pca")}

\item{legend.dot.size}{the size of the colored dots on your chart legend.
Default is 2}

\item{do.finetuning}{Performs the SingleR fine-tuning function.
Default is FALSE}

\item{local.celldex}{Provide a local copy of CellDex library.
Default is NULL}

\item{use.clusters}{Provide cluster identities for each cell.
Default is NULL}
}
\value{
a Seurat object with additional metadata
}
\description{
Annotate the cell types of your cells using SingleR (Aran et al., 2019). This
function takes a combined Seurat object after PC reduction and assigns cells
to a category (for example, stem cells or T cells) based on genomic profile.
}
\details{
This function is Step 5 of the basic Single-Cell RNA-seq workflow.
It is the starting point for downstream visualization, subsetting, and
analysis. It takes a combined seurat object as input, such as the one created
by the Combined&Renormalized function at the end of the Filter&QC Path
}
\examples{
\dontrun{
out <- annotateCellTypes(
  object = combined_so,
  species = "Human",
  reduction.type = "umap"
)
}


}
