% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Append_Metadata_to_Seurat_Objects.R
\name{appendMetadataToSeuratObject}
\alias{appendMetadataToSeuratObject}
\title{Append Metadata to Seurat Object.}
\usage{
appendMetadataToSeuratObject(object, metadata.to.append, sample.name.column)
}
\arguments{
\item{object}{The input Seurat Object or list of Seurat Objects to which you
wish to add metadata.}

\item{metadata.to.append}{A table of sample metadata that you want to append
to the already-existing metadata within the input Seurat Object(s).}

\item{sample.name.column}{The column of the input metadata.to.append table
that contains sample names matching the orig.idents in the input object(s).}
}
\value{
Function returns a Seurat Object or Objects with additional metadata
columns containing the appended metadata now annotated each cell by sample
name (orig.ident).
}
\description{
This template appends sample metadata from an input
table to a Seurat object, creating new metadata columns labeling all cells
from each sample by the new metadata values.
}
\details{
This template appends sample metadata from an input
table to a Seurat object, creating new metadata columns labeling all cells
from each sample by the new metadata values.
}
