% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_from_json}
\alias{cli_from_json}
\title{Call an SCWorkflow function with arguments specified in a json file}
\usage{
cli_from_json(method, json, debug = FALSE)
}
\arguments{
\item{method}{function in SCWorkflow to call}

\item{json}{path to a JSON file containing arguments for the function.
Additionally, the JSON can contain the following keys:
\itemize{
\item \code{object_input_rds} - filepath to an existing MultiOmicsDataset object in RDS format. This is required if the SCWorkflow function contains \code{object} as an argument.
\item \code{object_output_rds} - filepath to write the result to.
}}

\item{debug}{when TRUE, do not call the command, just return the expression.}
}
\value{
invisible returns the function call
}
\description{
Call an SCWorkflow function with arguments specified in a json file
}
\keyword{internal}
