% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_by_Genes_Automatic.R
\name{color_by_genes}
\alias{color_by_genes}
\title{Create gene expression annotated dimension plots based on user inputted table}
\usage{
color_by_genes(
  SO,
  samples_to_include,
  samples_to_display,
  marker_list,
  cells_of_interest,
  protein_presence = FALSE,
  assay = "SCT",
  reduction_type = "umap",
  point_transparency = 0.5,
  point_shape = 16,
  number_of_rows = 0,
  doCiteSeq = FALSE
)
}
\arguments{
\item{SO}{Seurat-class object}

\item{samples_to_include}{List of samples to subset the data by}

\item{samples_to_display}{List of samples to depict on dimension plot, samples not in the list would be colored gray in the background}

\item{marker_list}{Table of marker genes for each celltype (column names of the table), append "_prot" or "_neg" for proteins or negative markers}

\item{cells_of_interest}{Vector of celltypes from geneset_dataframe to screen for}

\item{protein_presence}{Set to TRUE if there are CITE-seq markers in geneset_dataframe}

\item{assay}{Name of the assay to extract gene expression data from}

\item{reduction_type}{Choose among tsne, umap, and pca}

\item{point_transparency}{Set to lower value for more see through points on dimension plot}

\item{point_shape}{Change the shape of points for between visualization}

\item{number_of_rows}{Set the number of rows to be displayed on the compiled gene expression dimension plot}

\item{doCiteSeq}{Set to TRUE if using Cite-seq data}
}
\value{
compiled dimension plots of markers, in the same layout as the user-inputted marker table
}
\description{
Returns a panel of dimension plots colored by individual gene expression
}
\details{
Takes in a list of genes inputted by the user, displays gene expression information on tsne, umap, or pca
}
