% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_by_Gene.R
\name{colorByGene}
\alias{colorByGene}
\title{Visualize gene expression for provided Genes
across your cells as a heatmap}
\usage{
colorByGene(
  object,
  samples.to.include,
  gene,
  reduction.type = "umap",
  number.of.rows = 0,
  return.seurat.object = FALSE,
  color = "red",
  point.size = 1,
  point.shape = 16,
  point.transparency = 0.5,
  use.cite.seq.data = FALSE,
  assay = "SCT"
)
}
\arguments{
\item{object}{Object of class Seurat}

\item{samples.to.include}{Samples to be included in the analysis}

\item{gene}{Genes which you would like to visualize}

\item{reduction.type}{Select the kind of clustering visualization
you would like to use to visualize the cell type results
("umap", "tsne", "pca"). Default is "umap"}

\item{number.of.rows}{The number of rows you want to arrange your plots into}

\item{return.seurat.object}{Set to FALSE if you want only a geneset
(and not a Seurat object) to be returned}

\item{color}{The color you want to use in your heatmap (default "red")}

\item{point.size}{The size of the points representing
each cell in your visualization. Default is 1}

\item{point.shape}{The code for your point shape (R "pch" argument).
Default is 16}

\item{point.transparency}{Set the transparency. Default is 0.5}

\item{use.cite.seq.data}{TRUE if you would like to plot Antibody clusters
from CITEseq instead of scRNA.}

\item{assay}{Select Assay to Plot (default is SCT)}
}
\value{
a Seurat object with additional metadata or gene table and plot
}
\description{
You should see one plot (TSNE or UMAP, your choice)
per gene name provided. The intensity of the red color will be relative
to the expression of the gene in that cell
}
\details{
This function must be run downstream of the Sample Names function,
as well as be provided a combined Seurat Object
such as the one produced by the SingleR Cell Annotation function
}
\examples{
\dontrun{
out <- colorByGene(
  object = anno_so,
  samples.to.include = c("sample1", "sample2"),
  gene = c("CD3D", "MS4A1"),
  reduction.type = "umap"
)
}

}
