% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Gene_Expression_Markers.R
\name{degGeneExpressionMarkers}
\alias{degGeneExpressionMarkers}
\title{DE with Find Markers \link{CCBR} \link{scRNA-seq}}
\usage{
degGeneExpressionMarkers(
  object,
  samples,
  contrasts,
  parameter.to.test = "orig_ident",
  test.to.use = "MAST",
  log.fc.threshold = 0.25,
  use.spark = FALSE,
  assay.to.use = "SCT"
)
}
\arguments{
\item{object}{Seurat-class object}

\item{samples}{Samples to be included in the analysis}

\item{contrasts}{Contrasts in the "A-B" format}

\item{parameter.to.test}{Select the metadata column that you would like
to use to perform your DEG analysis and construct your contrasts from.
Default is "orig_ident"}

\item{test.to.use}{The kind of algorithm you would like to use
to perform your DEG analysis. Default is the MAST algorithm
(wilcox,bimod,roc,t,negbinom,poisson,LR,MAST,DESeq2).}

\item{log.fc.threshold}{The minimum log fold-change between contrasts
that you would like to analyze. Default is 0.25}

\item{use.spark}{Opt to use Spark to parallelize computations.
Default is FALSE}

\item{assay.to.use}{The assay to use for your DEG analysis.
Default is SCT, but can use linearly scaled data by selecting RNA instead}
}
\value{
a dataframe with DEG.
}
\description{
This function performs DE (differential expression) analysis on
a merged Seurat object to identify expression markers between different
groups of cells (contrasts).
}
\details{
The recommended input is a merged Seurat object
with SingleR annotations, along with its associated sample names and metadata
}
\examples{
\dontrun{
deg <- degGeneExpressionMarkers(
  object = anno_so,
  samples = c("sample1", "sample2"),
  contrasts = c("A-B"),
  parameter.to.test = "cluster"
)
}



}
