% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_and_QC.R
\name{Filter_and_QC}
\alias{Filter_and_QC}
\title{Filter & QC Samples}
\usage{
Filter_and_QC(
  localFilePaths,
  organism = "Human",
  rename = FALSE,
  New_Sample_Names = c("Sample_1", "Sample_2", "Sample_3"),
  mincells = 3,
  mingenes = 200,
  complexity = 0.5,
  MAD_genes_value = 3,
  MAD_mitoch_value = NA,
  minUMI = 500,
  MAD_gene = TRUE,
  MAD_mitoch = TRUE,
  maxgenes = 2500,
  maxmitoch = 10,
  Filter_VDJ_Genes = FALSE,
  Keep = TRUE,
  File_Filter_Regex = "",
  Split_H5 = FALSE,
  Protein = FALSE,
  Cell_hash = FALSE,
  imageType = "png",
  plot_histogram = FALSE
)
}
\arguments{
\item{localFilePaths}{Please input your h5 files that have been uploaded to a single dataset. If you uploaded your h5 files through the Uploader page, this dataset should be the only node in the graph of your code workbook when you first launch a new analysis.}

\item{organism}{Please select what species your data comes from.}

\item{rename}{If FALSE, keep original sample names.  If TRUE, rename samples (see below to input your new sample names).}

\item{New_Sample_Names}{Input a list of new sample names to name your samples. Note: number of new names MUST match the number of samples. For reference, consult log output for Filter QC template for a list of old sample names in order.}

\item{mincells}{Filter out genes found in less than this number of cells. E.g. Setting to 3 will remove genes found in fewer than three cells of a sample.}

\item{mingenes}{Filter out cells with less than this number of genes found in them. E.g. setting to 200 will remove cells that have fewer than 200 genes from those analyzed for each sample.}

\item{complexity}{Number of genes detected per UMI. The more genes detected per UMI, the more complex the data. Cells that have a high number of UMIs but only a low number of genes could be dying cells, but also could represent a population of a low complexity cell type (i.e red blood cells). We suggest that you increase to 0.8 if samples have suspected RBC contamination.}

\item{MAD_genes_value}{How many Median Absolute Deviations do you want to use to filter out cells with too many genes? For example, entering "3" will remove all cells with 3 absolute deviations greater a number of genes than the median cell is calculated to have.}

\item{MAD_mitoch_value}{How many Median Absolute Deviations do you want to use to filter out cells with too high a percentage of mitochondrial RNA? For example, entering "3" will remove all cells with 3 absolute deviations greater a percent mitonchondrial content than the median cell is calculated to have.}

\item{minUMI}{Filter out cells with low number of reads}

\item{MAD_gene}{Filter by number of genes: If TRUE, uses median absolute deviation to detect outliers. If FALSE, uses an absolute threshold set below (see Filter maximum number of genes)}

\item{MAD_mitoch}{Filter by mitochondrial percentage: If TRUE, uses median absolute deviation to remove outliers. If FALSE, uses set value (below, Filter maximum percentage mitochondrial content)}

\item{maxgenes}{To remove potential doublets, set maximum number of genes per cell. E.g. Setting to 2,500 will remove cells with more than 2,500 genes.}

\item{maxmitoch}{Filter out cells whose proportion of mitochondrial genes exceed this threshold. E.g. setting to 10 removes cells with more than 10\% mitochondrial RNA.}

\item{Filter_VDJ_Genes}{Toggle off to remove VDJ genes from the scRNA transcriptome assay. This is to prevent clustering bias in T-cells of the same clonotype. Only recommended if you are also doing TCR-seq.}

\item{Keep}{TRUE is Keep, FALSE is Remove when pattern is found in the sample name. The pattern is set in the File Filter Regex parameter (above).}

\item{File_Filter_Regex}{Filter for pattern or regular expression in file name. Using this, you can filter on the name of the input files. Use the Keep or Remove parameter (below) to choose to keep or remove sample files whose names match this pattern.}

\item{Split_H5}{If TRUE, split H5 into individual files.  By default FALSE}

\item{Protein}{Does this dataset contain CITE-Seq antibodies? Set to TRUE if your dataset does contain CITE-Seq antibodies. Set to FALSE if your dataset does not. The default is FALSE.}

\item{Cell_hash}{Does this dataset contain Cell Hashing Tags? set to TRUE if your dataset does contain Cell Hashing Tags. Set to FALSE if your dataset does not. The default is FALSE.}

\item{imageType}{Remember that svgs are much larger than pngs, so we recommend doing everything first in png, then rerunning to output specific svgs as needed.}

\item{plot_histogram}{Set to TRUE to plot QC graphs as histograms. Keep as FALSE to plot as violin plots.}
}
\value{
Seurat Objects, Filters and Plots QC (before and after filtering) on samples. Takes a dataset of filtered h5 files (each representing one sample), and performs basic QC across several metrics. Also puts the data into a Seurat Object, the basic data structure for Seurat Single Cell analysis. This template is Step 1 of the basic Single-Cell RNA-seq workflow.
}
\description{
Creates Seurat Objects, Filters and Plots QC (before and after filtering) on samples. Takes a dataset of filtered h5 files (each representing one sample), and performs basic QC across several metrics. Also puts the data into a Seurat Object, the basic data structure for Seurat Single Cell analysis. This template is Step 1 of the basic Single-Cell RNA-seq workflow.
}
\details{
Filter and QC template in single-cell-rna-seq-r4 NIDAP environment
function from v 167
}
