% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony_Batch_Correction.R
\name{harmony_batch_correct}
\alias{harmony_batch_correct}
\title{Performs batch correction on Seurat-class object}
\usage{
harmony_batch_correct(
  so,
  variable_features = 2000,
  genes_to_add = c(),
  group.by.vars
)
}
\arguments{
\item{variable_features}{Number of most variable genes to subset the gene expression data by}

\item{genes_to_add}{Add genes that might not be found among variably expressed genes}

\item{group.by.vars}{which variable should be accountted for when running batch correction}

\item{seurat_object}{Seurat-class object}
}
\value{
Seurat-class Object with Harmony-adjusted gene expression (SCT slot) and tSNE cell embedding
}
\description{
Returns a Seurat-class object with adjusted cell embeddings and gene expression data to account for variations in sample collection
}
\details{
Takes in a list of genes inputted by the user, displays gene expression information in particular slot-assay with (optional) outliers removed
}
