% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony.R
\name{harmonyBatchCorrect}
\alias{harmonyBatchCorrect}
\title{Harmony Batch Correction from Singular Value Decomposed PCA}
\usage{
harmonyBatchCorrect(
  object,
  nvar = 2000,
  genes.to.add = c(),
  group.by.var,
  return.lognorm = T,
  npc = 30
)
}
\arguments{
\item{object}{Seurat-class object containing gene expression data and
metadata with the batch variable.}

\item{nvar}{Number of variable genes to subset the gene expression data by
(Default: 2000)}

\item{genes.to.add}{Add genes that might not be found among variably
expressed genes}

\item{group.by.var}{Which variable should be accounted for when running
batch correction}

\item{return.lognorm}{Logical; if TRUE, retain log-normalized assay behavior
in downstream handling. (Default: TRUE)}

\item{npc}{Number of principal components to use when running Harmony
(Default: 20)}
}
\value{
A list: adj.object with harmony-adjusted gene expression (SCT slot)
adj.tsne: harmonized tSNE plot
}
\description{
Adjusts cell embeddings and gene expression data to account for
variations due to user specified variable
}
\details{
Runs singular value decomposition on pearson residuals
(SCT scale.data) to obtain PCA embeddings. Performs harmony on
decomposed embedding and adjusts decomposed gene expression values
by harmonized embedding.
}
\examples{
\dontrun{
harmonyBatchCorrect(
  object = seurat,
  nvar = 2000,
  genes.to.add = c("Cd4", "Cd8a"),
  group.by.var = "Mouse_Origin",
  npc = 20
)
}
}
