% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{Plot coexpression of 2 markers using transcript and/or protein expression values}
\usage{
Heatmap(
  object,
  sample.names,
  metadata,
  transcripts,
  proteins = NULL,
  plot.title = "Heatmap",
  add.gene.or.protein = TRUE,
  protein.annotations = NULL,
  rna.annotations = NULL,
  arrange.by.metadata = TRUE,
  add.row.names = TRUE,
  add.column.names = FALSE,
  set.seed = 6,
  scale.data = TRUE,
  trim.outliers = TRUE,
  trim.outliers.percentage = 0.01,
  order.heatmap.rows = FALSE,
  row.order = c()
)
}
\arguments{
\item{object}{Seurat-class object}

\item{sample.names}{Sample names}

\item{metadata}{Metadata column to plot}

\item{transcripts}{Transcripts to plot}

\item{proteins}{Proteins to plot (default is NULL)}

\item{plot.title}{Title of plot (default is "Heatmap")}

\item{add.gene.or.protein}{Add Gene or protein annotations (default is FALSE)}

\item{protein.annotations}{Protein annotations to add (defulat is NULL)}

\item{rna.annotations}{Gene annotations to add (default is NULL)}

\item{arrange.by.metadata}{Arrange by metadata (default is TRUE)}

\item{add.row.names}{Add row names (default is FALSE)}

\item{add.column.names}{Add column names (default is FALSE)}

\item{set.seed}{Seed for colors (default is 6)}

\item{scale.data}{Perform z-scaling on rows (default is TRUE)}

\item{trim.outliers}{Remove outlier data (default is TRUE)}

\item{trim.outliers.percentage}{Set outlier percentage (default is 0.01)}

\item{order.heatmap.rows}{Order heatmap rows (default is FALSE)}

\item{row.order}{Set row order to gene vector (default is NULL)}
}
\description{
Returns individual and combined expression of 2 genes and allows for filtering (optional) of the Seurat object using expression thresholds
}
\details{
This method provides visualization of coexpression of 2 genes (or proteins) and additional methods for filtering for cells with gene expression values that are above or below thresholds set for one or both markers.
}
