% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleScoreApp.R
\name{launch_module_score_app}
\alias{launch_module_score_app}
\title{Launch the ModuleScore Shiny App}
\usage{
launch_module_score_app()
}
\value{
The result of \code{shiny::runApp()}
}
\description{
Opens the interactive app to explore ModuleScores and adjust thresholds.
}
