% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_Table.R
\name{MetadataTable}
\alias{MetadataTable}
\title{Extracting meta.data and reductions slots from Seurat Object}
\usage{
MetadataTable(SO, return.cell.embeddings = FALSE)
}
\arguments{
\item{SO}{object of class Seurat}

\item{return.cell.embeddings}{If TRUE, cell embeddings from dimensional reductions present in the Seurat Object are added to the output metadata table}
}
\value{
a data.frame extracted from the Seurat object
}
\description{
Returns the per cell metadata and dimension embeddings (optional) of the Seurat Object as a data table.
}
\details{
Exposes the metadata of any combined Seurat object, though it is recommended to input a SingleR annotated Seurat Object.
}
\examples{
\dontrun{ 
# Return metadata table
metadata <- MetadataTable(SO = seurat.object)
head(metadata)

# Return dimensional reductions
metadata <- MetadataTable(SO = seurat.object, return.cell.embeddings = TRUE)
umap_coordinates <- metadata \%>\% dplyr::select(Barcode, seurat_clusters, contains("UMAP"))

## Plot 
library(ggplot2)
ggplot(umap_coordinates, aes(x = UMAP_1, y = UMAP_2, colour = seurat_clusters)) + 
geom_point(size = 0.1) + theme_bw()
}

}
\seealso{
\code{\link[=SampleNames]{SampleNames()}} \code{\link[Seurat:reexports]{Seurat::AddMetaData()}}
}
