% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleScoreHelpers_011726.R
\name{modscore-imports-011726}
\alias{modscore-imports-011726}
\title{Helpers for ModuleScore Shiny app}
\description{
Precompute module scores per celltype and build plots from cached data.
}
\keyword{internal}
