% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleScore.R
\name{ModuleScore}
\alias{ModuleScore}
\title{Calculate average gene expression (subtracted by aggregated control feature set) and label cells according to ModuleScores}
\usage{
ModuleScore(
  SO,
  sample_names,
  sample_to_display,
  geneset_dataframe,
  proteins_presence = FALSE,
  celltypes_to_analyze,
  manual_threshold = c(0),
  general_class,
  multi_level_class = FALSE,
  levels_dataframe,
  reduction = "tsne",
  nbins = 24,
  gradient_density_font_size = 6,
  violinplot_font_size = 6,
  step_size = 0.1
)
}
\arguments{
\item{SO}{Seurat-class object}

\item{sample_names}{List of samples to subset the data by}

\item{sample_to_display}{List of samples to depict on dimension plot, samples not in the list would be colored gray in the background}

\item{geneset_dataframe}{Table of marker genes for each celltype (column names of the table), append "_prot" or "_neg" for proteins or negative markers}

\item{proteins_presence}{Set to TRUE if there are CITE-seq markers in geneset_dataframe}

\item{celltypes_to_analyze}{Vector of celltypes from geneset_dataframe to screen for}

\item{manual_threshold}{Specify bimodal thresholds for cell classification, should be of the same length as celltypes_to_analyze vector}

\item{general_class}{Base population of cells to classify}

\item{multi_level_class}{Toggle to TRUE if there are subpopulations of cells you want to screen for}

\item{levels_dataframe}{Table of subpopulation levels (column names of the table) and parent-child information (e.g. Tcells-CD4, Tcells-CD8)}

\item{reduction}{Choose among tsne, umap, and pca}

\item{nbins}{Number of bins for storing control features and analyzing average expression}

\item{gradient_density_font_size}{Set size of axis labels on gradient density plot of ModuleScore distribution}

\item{violinplot_font_size}{Set size of axis labels on violin plot of ModuleScore distribution}

\item{step_size}{Set step size of distribution plots}

\item{image_width}{Set Image width (pixel units)}

\item{image_height}{Set Image height (pixel units)}

\item{image_type}{Choose between png and svg}
}
\value{
Annotated dimension plot with ModuleScore distribution of cell marker gene, Seurat Object with cell classification metadata
}
\description{
Returns Seurat-class object with metadata containing ModuleScores and Likely_CellType calls
}
\details{
Analyzed features are binned based on averaged expression; control features are randomly selected from each bin.
}
