% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palantir_api_call.R
\name{palantir_api_call}
\alias{palantir_api_call}
\title{palantir_api_call
Utility function from 3D tSNE Coordinate Template from v 75#'}
\usage{
palantir_api_call(service, path, token, data, method)
}
\arguments{
\item{service}{The NIDAP API service to call}

\item{path}{The path of NIDAP API service}

\item{token}{NIDAP user toekn.}

\item{data}{Data to be uploaded with NIDAP API calls.}

\item{method}{Method to be used, including POST, GET, and DELETE}
}
\value{
return the content of API calls
}
\description{
palantir_api_call
Utility function from 3D tSNE Coordinate Template from v 75#'
}
