% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_and_Normalization.R
\name{PCA_and_Normalization}
\alias{PCA_and_Normalization}
\title{PCA & Normalization}
\usage{
PCA_and_Normalization(
  Seurat_Object,
  vars_to_regress = NULL,
  vars_to_plot = c("percent.mt", "Phase"),
  npcs = 30,
  nfeatures = 2000,
  low_cut = 1,
  high_cut = 8,
  low_cut_disp = 1,
  high_cut_disp = 1e+05,
  selection_method = "vst",
  doJackStraw = FALSE,
  JackStraw_dims = 5,
  methods_PCA = c("Elbow", "Marchenko-Pastur"),
  var_threshold = 0.1,
  imageType = "png"
)
}
\arguments{
\item{Seurat_Object}{Please input a post-Filter QC Seurat Object to do your reduction on.}

\item{vars_to_regress}{Subtract (‘regress out’) this source of heterogeneity from the data. For example, to regress out mitochondrial effects, input "percent.mt." Options: percent.mt, nCount_RNA, S.Score, G2M.Score, CC.Difference}

\item{vars_to_plot}{Choose variables to visualize on PCA plot (maximum = 3)}

\item{npcs}{Select initial set of principal components}

\item{nfeatures}{Number of variable features}

\item{low_cut}{description.}

\item{high_cut}{description.}

\item{low_cut_disp}{description.}

\item{high_cut_disp}{description.}

\item{selection_method}{Method to choose top variable features. Options: vst, mean.var.plot, dispersion}

\item{doJackStraw}{Opt to visualize your data in a Jackstraw plot (sometimes more description than an elbow plot).}

\item{JackStraw_dims}{Recommended max 10}

\item{methods_PCA}{Methods available: Marchenko-Pastur: use eigenvalue null upper bound from URD, Elbow: Find threshold where percent change in variation between consecutive PCs is less than X\% (set below)}

\item{var_threshold}{For Elbow method, set percent change threshold in variation between consecutive PCs}

\item{imageType}{Remember that svgs are much larger than pngs, so we recommend doing everything first in png, then rerunning to output specific svgs as needed. Options: png, svg}
}
\value{
Seurat Objects and QC plots. This template normalizes each sample and plots Seurat PCAs and Elbow Plots for each sample to aid in selecting the number of PCs to carry forward in your analysis (see next template in workflow). PCA plots before and after regression will be shown for each variable chosen. The elbow plot will show reduction in variance (y-axis) as a function of PC count (x-axis). It is recommended to pick the number of PCs where variance is reduced by 98\% (i.e. the first point where y coord = 0.02). This is Step 3 in the basic Single-Cell RNA-seq workflow.
}
\description{
This template normalizes each sample and plots Seurat PCAs and Elbow Plots for each sample to aid in selecting the number of PCs to carry forward in your analysis (see next template in workflow). PCA plots before and after regression will be shown for each variable chosen. The elbow plot will show reduction in variance (y-axis) as a function of PC count (x-axis). It is recommended to pick the number of PCs where variance is reduced by 98\% (i.e. the first point where y coord = 0.02). This is Step 3 in the basic Single-Cell RNA-seq workflow.
}
\details{
PCA & Normalization template in single-cell-rna-seq-r4 NIDAP environment
from v96
}
