% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulk_DEG.R
\name{Pseudobulk_DEG}
\alias{Pseudobulk_DEG}
\title{Pseudobulk Differential Gene Expression Analysis}
\usage{
Pseudobulk_DEG(
  so,
  contrasts,
  replicate,
  comparison_level,
  label,
  min_num_cells = 3,
  min_num_reps = 1,
  min_num_features = 0,
  subdivide_data = FALSE
)
}
\arguments{
\item{so}{Seurat-class object}

\item{contrasts}{Choose the groups (group1-group2) to compare transcription profiles against, positive Fold Change values indicate upregulation in group1}

\item{replicate}{Column of metadata containing replicate information, gene expression will first be aggregated based on this column before differential analysis is ran}

\item{comparison_level}{Name of metadata column containing information on the sub-levels you want to divide your data into.}

\item{label}{Name of metadata column containing information on the groups (contrasts) you want to compare.}

\item{min_num_cells}{Filters a comparison level based on the minimum number of cells at that level.}

\item{min_num_reps}{Filters a comparison level based on the minimum number of replicates at that level.}

\item{min_num_features}{Filters out genes if they are expressed in fewer than X number of cells.}

\item{subdivide_data}{Toggle to TRUE if you want to make comparisons across different levels within your data.}
}
\value{
data.frame of differentially expressed genes across comparisons
}
\description{
Performs Linear Models for Microarray Data on single-cell data
}
\details{
Aggregate gene expression amongst replicates before performing limma regression
}
