% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Names.R
\name{SampleNames}
\alias{SampleNames}
\title{Returns tabulation of cell identities present in the orig_ident column}
\usage{
SampleNames(SO)
}
\arguments{
\item{SO}{Seurat-class object}
}
\value{
A table (data.frame) with the number of cells per each sample name (column names)
}
\description{
It is recommended to run this function after doing SingleR annotations and before doing any visualizations/downstream analysis that requires inputting sample names.
}
