% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tSNE.R
\name{tSNE3D}
\alias{tSNE3D}
\title{Plot 3D-TSNE given a Seurat Object and returns plotly image}
\usage{
tSNE3D(
  object,
  color.variable,
  label.variable,
  dot.size = 4,
  legend = TRUE,
  colors = c("darkblue", "purple4", "green", "red", "darkcyan", "magenta2", "orange",
    "yellow", "black"),
  filename = "plot.html",
  save.plot = FALSE,
  npcs = 15
)
}
\arguments{
\item{object}{Seurat-class object}

\item{color.variable}{Metadata column in Seurat Object to use for color}

\item{label.variable}{Metadata column in Seurat Object to use for label}

\item{dot.size}{Dot size for plot (default is 4)}

\item{legend}{If TRUE, show legend (default is TRUE)}

\item{colors}{Colors used for the color.variable}

\item{filename}{Filename for saving plot (default is "plot.html")}

\item{save.plot}{Save plot as widget in html file (default is FALSE)}

\item{npcs}{Number of principal components used for tSNE calculations
(default is 15)}
}
\value{
A list with a plotly 3D TSNE plot (\code{figure}) and TSNE coordinates
(\code{tsne.df}).
}
\description{
This method provides visualization of 3D-tSNE plot given a
Seurat Object and returns a plotly plot and a dataframe of TSNE
coordinates. It optionally saves the plotly image embedded in an html file.
}
\examples{
\dontrun{
out <- tSNE3D(
  object = seurat_obj,
  color.variable = "cell_type",
  label.variable = "orig.ident",
  npcs = 15,
  save.plot = FALSE
)
}
}
