% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_by_Genes_Automatic.R
\name{colorByMarkerTable}
\alias{colorByMarkerTable}
\title{Color by Gene List}
\usage{
colorByMarkerTable(
  object,
  samples.subset,
  samples.to.display,
  marker.table,
  cells.of.interest,
  protein.presence = FALSE,
  assay = "SCT",
  reduction.type = "umap",
  point.transparency = 0.5,
  point.shape = 16,
  cite.seq = FALSE
)
}
\arguments{
\item{object}{Seurat-class object}

\item{samples.subset}{List of samples to subset the data by}

\item{samples.to.display}{List of samples to depict on dimension plot,
samples not in the list would be colored gray in
the background}

\item{marker.table}{Table of marker genes for each celltype
(column names of the table), append "_prot" or "_neg" for
proteins or negative markers}

\item{cells.of.interest}{Celltypes from geneset_dataframe to screen for}

\item{protein.presence}{Set to TRUE if protein markers are used}

\item{assay}{Assay to extract gene expression data from (Default: "SCT")}

\item{reduction.type}{Choose among tsne, umap, and pca (Default: "umap")}

\item{point.transparency}{Set to lower values for more see through points on
dimension plot (Default: 0.5)}

\item{point.shape}{Change the shape of points for between visualization
(Default: 16)}

\item{cite.seq}{Set to TRUE to use CITE-seq embedding for dimension reduction}
}
\value{
arranged grob of dimension reduction plots colored by individual
marker expression
}
\description{
Returns a panel of reduction plots colored by marker expression
}
\details{
Takes in a gene table inputted by the user, displays a panel of
tsne, umap, or pca colored by marker expression. The panel will be
organized in a similar format as the gene table, but with the
omission of genes not found in the data
}
