% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dotplot_by_Metadata.R
\name{DotplotMet}
\alias{DotplotMet}
\title{Dotplot of Gene Expression by Metadata}
\usage{
DotplotMet(
  object,
  metadata,
  cells,
  markers,
  plot.reverse = FALSE,
  cell.reverse.sort = FALSE,
  dot.color = "darkblue"
)
}
\arguments{
\item{metadata}{Metadata category to plot}

\item{cells}{Metadata category factors to plot}

\item{markers}{Column that contains gene names}

\item{plot.reverse}{Set Metadata categories to x-axis}

\item{cell.reverse.sort}{Reverse Metadata Categories to plot}

\item{dot.color}{Dot color (default is "dark blue")}

\item{input.dataset}{Seurat Object}
}
\value{
Dotplot with markers and cell types.
}
\description{
This Dotplot plotter plots average gene expression values for a set of genes from an input table.  Input table contains a single column for "Genes" and a single column for "Cell Type".  The values in the "Cell Type" column should match the values provided in the metadata template (Metadata Category to Plot).  The plot will order the genes (x-axis, left to right) and Cell Types (y-axis, top to bottom) in the order in which it appears in the input table.  Based on the additional column selected (Sample Column where sample or any other metadata category), it will display a contingency table for the Metadata
}
\details{
This method provides a visualization plot showing the frequency of positively
}
