% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dotplot_by_Metadata.R
\name{DotplotMet}
\alias{DotplotMet}
\title{Dotplot of Gene Expression by Metadata}
\usage{
DotplotMet(
  object,
  metadata,
  cells,
  markers,
  plot.reverse = FALSE,
  cell.reverse.sort = FALSE,
  dot.color = "darkblue"
)
}
\arguments{
\item{object}{Seurat Object}

\item{metadata}{Metadata category to plot}

\item{cells}{Metadata category factors to plot}

\item{markers}{Column that contains gene names}

\item{plot.reverse}{Set Metadata categories to x-axis (default is FALSE)}

\item{cell.reverse.sort}{Reverse Metadata Categories to plot (default is FALSE)}

\item{dot.color}{Dot color (default is "dark blue")}
}
\value{
Dotplot with markers and cell types.
}
\description{
This function uses the Dotplot function from Seurat and plots
average gene expression values and percent expressed for a set of genes.
}
\details{
This method provides a dotplot showing the percent frequency of
gene-positive cells as size of dot and degree of expression as color
of dot.
}
