% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dotplot_by_Metadata.R
\name{dotPlotMet}
\alias{dotPlotMet}
\title{Dotplot of Gene Expression by Metadata}
\usage{
dotPlotMet(
  object,
  metadata,
  cells,
  markers,
  plot.reverse = FALSE,
  cell.reverse.sort = FALSE,
  dot.color = "darkblue"
)
}
\arguments{
\item{object}{Seurat Object}

\item{metadata}{Metadata column in Seurat Object to plot}

\item{cells}{Vector of metadata category factors to plot and should be found
in metadata column. Order of plotting will follow exact order as entered.}

\item{markers}{Vector of genes to plot.  Order of plotting will follow exact
order as entered}

\item{plot.reverse}{If TRUE, set metadata categories to x-axis and genes to
y-axis (default is FALSE)}

\item{cell.reverse.sort}{If TRUE, Reverse plot order of metadata category
factors (default is FALSE)}

\item{dot.color}{Dot color (default is "dark blue")}
}
\value{
Dotplot with markers and cell types.
}
\description{
This function uses the Dotplot function from Seurat and plots
average gene expression values and percent expressed for a set of
genes.
}
\details{
This method provides a dotplot showing the percent frequency of
gene-positive cells as size of dot and degree of expression as color
of dot.
}
