% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dual_Labeling.R
\name{dualLabeling}
\alias{dualLabeling}
\title{Cell Annotation with Co-Expression \link{CCBR} \link{scRNA-seq}}
\usage{
dualLabeling(
  object,
  samples,
  marker.1,
  marker.2,
  marker.1.type = "SCT",
  marker.2.type = "SCT",
  data.reduction = "both",
  point.size = 0.5,
  point.shape = 16,
  point.transparency = 0.5,
  add.marker.thresholds = TRUE,
  marker.1.threshold = 0.5,
  marker.2.threshold = 0.5,
  filter.data = TRUE,
  marker.1.filter.direction = "greater than",
  marker.2.filter.direction = "greater than",
  apply.filter.1 = TRUE,
  apply.filter.2 = TRUE,
  filter.condition = TRUE,
  parameter.name = "My_CoExp",
  trim.marker.1 = FALSE,
  trim.marker.2 = FALSE,
  pre.scale.trim = 0.99,
  display.unscaled.values = FALSE
)
}
\arguments{
\item{object}{Seurat-class object}

\item{samples}{Samples to be included in the analysis}

\item{marker.1}{First gene/marker for coexpression analysis}

\item{marker.2}{Second gene/marker for coexpression analysis}

\item{marker.1.type}{Slot to use for first marker. Choices are "SCT",
"protein","HTO", or "Spatial" (default is "SCT")}

\item{marker.2.type}{Slot to use for second marker. Choices are "SCT",
"protein","HTO", or "Spatial" (default is "SCT")}

\item{data.reduction}{Dimension Reduction method to use for image. Options
are "umap", "tsne", or "both" (default is "both")}

\item{point.size}{Point size for image (default is 0.5)}

\item{point.shape}{Point shape for image (default is 16)}

\item{point.transparency}{Point transparency for image (default is 0.5)}

\item{add.marker.thresholds}{Add marker thresholds on plot (default is TRUE)}

\item{marker.1.threshold}{Threshold set for first marker (default is 0.5)}

\item{marker.2.threshold}{Threshold set for second marker (default is 0.5)}

\item{filter.data}{Add new parameter column to metadata annotating where
marker thresholds are applied (default is TRUE)}

\item{marker.1.filter.direction}{Annotate cells that have gene expression levels
for marker 1 using the marker 1 threshold. Choices are "greater than"
or "less than" (default is "greater than")}

\item{marker.2.filter.direction}{Annotate cells that have gene expression levels
for marker 2 using the marker 2 threshold. Choices are "greater than"
or "less than" (default is "greater than")}

\item{apply.filter.1}{If TRUE, apply the first filter (default is TRUE)}

\item{apply.filter.2}{If TRUE, apply the second filter (default is TRUE)}

\item{filter.condition}{If TRUE, apply both filters 1 and 2 and take
intersection. If FALSE, apply both filters and take the union.}

\item{parameter.name}{Name for metadata column for new marker filters
(default is "My_CoExp")}

\item{trim.marker.1}{Trim top and bottom percentile of marker 1 signal to
pre-scale trim values (below) to remove extremely low and high values
(default is FALSE)}

\item{trim.marker.2}{Trim top and bottom percentile of marker 2 signal to
pre-scale trim values (below) to remove extremely low and high values
(default is FALSE)}

\item{pre.scale.trim}{Set trimming percentile value (default is 0.99)}

\item{display.unscaled.values}{Set to TRUE if you want to view the unscaled
gene/protein expression values (default is FALSE)}
}
\value{
a seurat object with optional additional metadata for cells that are
positive or negative for gene markers, a coexpression plot and contingency
table showing sum of cells filtered.
}
\description{
Display co-expression of two chosen markers in your Seurat
object. Creates a metadata column containing annotations for cells that
correspond to marker expression thresholds.
}
\examples{
\dontrun{
out <- dualLabeling(
  object = anno_so,
  samples = c("sample1"),
  marker.1 = "CD3D",
  marker.2 = "MS4A1",
  data.reduction = "umap"
)
}

}
