% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dual_Labeling.R
\name{DualLabeling}
\alias{DualLabeling}
\title{Plot coexpression of 2 markers using transcript and/or protein expression values}
\usage{
DualLabeling(
  object,
  samples,
  marker1,
  marker_1_type,
  marker2,
  marker_2_type,
  data_reduction = "umap",
  point_size = 0.5,
  point_shape = 16,
  point_transparency = 0.5,
  add_marker_thresholds = FALSE,
  marker_1_threshold = 0.5,
  marker_2_threshold = 0.5,
  filter_data = FALSE,
  M1_filter_direction = "greater than",
  M2_filter_direction = "greater than",
  apply_filter_1 = TRUE,
  apply_filter_2 = TRUE,
  filter_condition = TRUE,
  parameter_name = "Marker",
  trim_marker_1 = TRUE,
  trim_marker_2 = TRUE,
  pre_scale_trim = 0.99,
  density_heatmap = FALSE,
  display_unscaled_values = FALSE
)
}
\arguments{
\item{object}{Seurat-class object}

\item{samples}{Samples to be included in the analysis}

\item{marker1}{First gene/marker for coexpression analysis}

\item{marker_1_type}{Slot to use for first marker (choices are "SCT","protein","HTO")}

\item{marker2}{Second gene/marker for coexpression analysis}

\item{marker_2_type}{Slot to use for second marker (choices are "SCT","protein","HTO")}

\item{data_reduction}{Dimension Reduction method to use for image (options are "umap" or "tsne")}

\item{point_size}{Point size for image (default is 0.5)}

\item{point_shape}{Point shape for image (default is 16)}

\item{point_transparency}{Point transparency for image (default is 0.5)}

\item{add_marker_thresholds}{Add marker thresholds (default is FALSE)}

\item{marker_1_threshold}{Threshold set for first marker (default is 0.5)}

\item{marker_2_threshold}{Threshold set for first marker (default is 0.5)}

\item{filter_data}{Add new parameter to metadata using marker thresholds (default is FALSE)}

\item{M1_filter_direction}{Filter to samples that have greater or less than marker 1 threshold (default is "greater than")}

\item{M2_filter_direction}{Filter to samples that have greater or less than marker 2 threshold (default is "greater than")}

\item{apply_filter_1}{If TRUE, apply the first filter (default is TRUE)}

\item{apply_filter_2}{If TRUE, apply the second filter (default is TRUE)}

\item{filter_condition}{If TRUE, apply both filters 1 and 2 and take intersection. If FALSE, apply filters and take the union.}

\item{parameter_name}{Name for metadata column for new marker filters. (Default is "Marker")}

\item{trim_marker_1}{Trim top and bottom percentile of marker 1 signal to pre-scale trim values (below) to remove extremely low and high values (Default is TRUE)}

\item{trim_marker_2}{Trim top and bottom percentile of marker 2 signal to pre-scale trim values (below) to remove extremely low and high values (Default is TRUE)}

\item{pre_scale_trim}{Set trimming percentile values (Defalut is 0.99)}

\item{density_heatmap}{Creates a additional heatmap showing the density distribution of cells. (Default is FALSE)}

\item{display_unscaled_values}{Set to TRUE if you want to view the unscaled gene/protein expression values (Default is FALSE)}
}
\value{
a seurat object with optional additional metadata for cells that are positive or negative for gene markers and a coexpression plot.
}
\description{
Returns individual and combined expression of 2 genes and allows for filtering (optional) of the Seurat object using expression thresholds
}
\details{
This method provides visualization of coexpression of 2 genes (or proteins) and additional methods for filtering for cells with gene expression values that are above or below thresholds set for one or both markers.
}
