% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_QC.R
\name{filterQC}
\alias{filterQC}
\title{Filter Low Quality Cells \link{CCBR} \link{scRNA-seq}}
\usage{
filterQC(
  object,
  min.cells = 20,
  filter.vdj.genes = F,
  nfeature.limits = c(NA, NA),
  mad.nfeature.limits = c(5, 5),
  ncounts.limits = c(NA, NA),
  mad.ncounts.limits = c(5, 5),
  mitoch.limits = c(NA, 25),
  mad.mitoch.limits = c(NA, 3),
  complexity.limits = c(NA, NA),
  mad.complexity.limits = c(5, NA),
  topNgenes.limits = c(NA, NA),
  mad.topNgenes.limits = c(5, 5),
  n.topgenes = 20,
  do.doublets.filter = TRUE,
  plot.outliers = "None",
  group.column = NA,
  nfeatures = 2000,
  low.cut = 0.1,
  high.cut = 8,
  low.cut.disp = 1,
  high.cut.disp = 1e+05,
  selection.method = "vst",
  npcs = 30,
  vars.to.regress = NULL,
  seed.for.PCA = 42,
  seed.for.TSNE = 1,
  seed.for.UMAP = 42
)
}
\arguments{
\item{object}{a list of seurat objects for each sample.}

\item{min.cells}{Filter out genes found in less than this number of cells.
E.g. Setting to 20 will remove genes found in fewer than 3 cells of
a sample. (Default: 20)}

\item{filter.vdj.genes}{If FALSE to remove VDJ genes from the scRNA
transcriptome assay. This is to prevent clustering bias in T-cells of the
same clonotype. Only recommended if you are also doing TCR-seq.
(Default: FALSE)}

\item{nfeature.limits}{Filter out cells where the number of genes found in each
cell exceed the selected lower or upper limits.
Usage c(lower limit, Upper Limit). E.g. setting to c(200,1000) will remove
cells that have fewer than 200 genes or more than 1000 genes
for each sample. (Default: c(NA, NA))}

\item{mad.nfeature.limits}{Set filter limits based on how many Median
Absolute Deviations an outlier cell will have. Calculated from the median
gene number for all cells in your sample. Usage c(lower limit, Upper Limit)
E.g. setting to c(3,5) will remove all cells with more than 3 absolute
deviations less than the median or 5 absolute deviations greater than the
median. (Default: c(5,5))}

\item{ncounts.limits}{Filter out cells where the total number of molecules
(umi) detected within a cell exceed the selected limits.
Usage c(lower limit, Upper Limit). E.g. setting to c(200,100000) will remove
cells that have fewer than 200 or greater than 100000 molecules.
(Default: c(NA, NA))}

\item{mad.ncounts.limits}{Set filter limits based on how many Median Absolute
Deviations an outlier cell will have. Calculated from the median number of
molecules for all cells in your sample. Usage c(lower limit, Upper Limit)
E.g. setting to c(3,5) will remove all cells with more than 3 absolute
deviations less than the median or with more than 5 absolute deviations
greater than the median. (Default: c(5,5))}

\item{mitoch.limits}{Filter out cells whose proportion of mitochondrial genes
exceed the selected lower or upper limits.
Usage c(lower limit, Upper Limit). E.g. setting to c(0,8) will not set the
lower limit and removes cells with more than 8\% mitochondrial RNA.
(Default: c(NA,25))}

\item{mad.mitoch.limits}{Set filter limits based on how many Median Absolute
Deviations an outlier cell will have. Calculated from the Median percentage
of mitochondrial RNA for all cells in your sample.
Usage c(lower limit, Upper Limit). E.g. setting to c(NA,3) will not set a
lower limit and remove all cells with more than 3 absolute deviations
greater than the median. (Default: c(NA,3))}

\item{complexity.limits}{Complexity represents Number of genes detected per
UMI. The more genes detected per UMI, the more complex the data.
Filter out cells whose Complexity exceed the selected lower or upper limits.
Cells that have a high number of UMIs but only a low number of genes could
be dying cells, but also could represent a population of a low complexity
cell type (i.e red blood cells). We suggest that you set the lower limit to
0.8 if samples have suspected RBC contamination.
Usage c(lower limit, Upper Limit). E.g. setting to c(0.8,0) will not set
an upper limit and removes cells with complexity less than 0.8.
(Default: c(NA,NA))}

\item{mad.complexity.limits}{Set filter limits based on how many Median
Absolute Deviations an outlier cell will have. Calculated from the Median
complexity for all cells in your sample.
Usage c(lower limit, Upper Limit). E.g. setting to c(5,NA) will not set an
upper limit and remove all cells with  more than 5 absolute deviations
less than the median. (Default: c(5,NA))}

\item{topNgenes.limits}{Filter Cells based on the percentage of total counts
in top N most highly expressed genes. Outlier cells would have a high
percentage of counts in just a few genes and should be removed.
The same considerations outlined in "complexity.limits" should be taken for
this filter.
Usage c(lower limit, Upper Limit). E.g. setting to c(NA,50) will not set a
lower limit and remove cells with greater than 50\% of reads in the top N
genes. (Default: c(NA,NA))}

\item{mad.topNgenes.limits}{Filter limits based on how many Median
Absolute Deviations an outlier cell will have. Calculated from the Median
percentage of counts in the top N Genes.
Usage c(lower limit, Upper Limit). E.g. setting to c(5,5) will remove all
cells with more than 5 absolute deviations greater than or 5 absolute
deviations less than the median percentage. (Default: c(5,5))}

\item{n.topgenes}{Select the number of top highly expressed genes used to
calculate the percentage of reads found in these genes.
E.g. a value of 20 calculates the percentage of reads found in the top 20
most highly expressed Genes.
(Default: 20)}

\item{do.doublets.filter}{Use scDblFinder to identify and remove doublet
cells. Doublets are defined as two cells that are sequenced under the same
cellular barcode, for example, if they were captured in the same droplet.
(Default: TRUE)}
}
\value{
Seurat Object and QC plots
}
\description{
Filters cells and genes across various criteria for each sample.
Multiple cell and gene filters can be selected to remove poor quality data
and noise while generating QC plots before and after filtering.
}
\details{
This is Step 2 in the basic Single-Cell RNA-seq workflow. Multiple
cell and gene filters can be selected to remove poor quality data and noise.
Workflows can use this downstream of any Seurat Object. This tool is
typically the second step in the Single Cell Workflow.
}
\examples{
\dontrun{
out <- filterQC(
  object = so_list,
  min.cells = 20,
  n.topgenes = 20,
  do.doublets.filter = TRUE
)
}

}
