% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony.R
\name{harmonyBatchCorrect}
\alias{harmonyBatchCorrect}
\title{Harmony Batch Correction from Singular Value Decomposed PCA}
\usage{
harmonyBatchCorrect(
  object,
  nvar = 2000,
  genes.to.add = c(),
  group.by.var,
  npc = 20
)
}
\arguments{
\item{nvar}{Number of variable genes to subset the gene expression data by
(Default: 2000)}

\item{genes.to.add}{Add genes that might not be found among variably
expressed genes}

\item{group.by.var}{Which variable should be accounted for when running
batch correction}

\item{npc}{Number of principal components to use when running Harmony
(Default: 20)}

\item{seurat_object}{Seurat-class object}
}
\value{
A list: adj.object with harmony-adjusted gene expression (SCT slot)
adj.tsne: harmonized tSNE plot
}
\description{
Adjusts cell embeddings and gene expression data to account for
variations due to user specified variable
}
\details{
Runs singular value decomposition on pearson residuals
(SCT scale.data) to obtain PCA embeddings. Performs harmony on
decomposed embedding and adjusts decomposed gene expression values
by harmonized embedding.
}
