% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleScore.R
\name{modScore}
\alias{modScore}
\title{Compute ModScore}
\usage{
modScore(
  object,
  marker.table,
  group.var = "orig.ident",
  use.columns,
  ms.threshold,
  general.class,
  multi.lvl = FALSE,
  lvl.df = NULL,
  reduction = "tsne",
  nbins = 10,
  gradient.ft.size = 6,
  violin.ft.size = 6,
  step.size = 0.1
)
}
\arguments{
\item{object}{Seurat-class object}

\item{marker.table}{A table of lists of gene/protein markers for the
categories of cells you want to detect.
The table should be formatted with cell type(s)
as the column names, and marker(s) as the entries
in each column.
Requires SCT@data to be present within Seurat Object}

\item{group.var}{Metadata column used for grouping in diagnostic plots.
(Default: "orig.ident")}

\item{use.columns}{Select specific columns within Marker Table to analyze.
Markers from unselected columns won't be included.}

\item{ms.threshold}{Allow user-specified module score thresholds.
Provide one threshold for each Celltype you included
in the "use.columns" parameter.
For each Celltype, provide the Celltype name,
then a space, then type your threshold for that Celltype.
This threshold must be a number between 0.0 and 1.0.
E.g. "Tcells 0.2", "Macrophages 0.37".
For best results, follow these steps:
(1) Set thresholds at 0.0 for a preliminary view of data.
(2) Use the resulting visualizations to estimate the
correct thresholds for each Celltype.
(3) Adjust your thresholds based on what you saw in the
visualizations.
(4) Re-run the template with the new thresholds.
(5) Review the visualizations again and
repeat Steps 1-5 if you think thresholds can be
further improved.}

\item{general.class}{Select which of the classes (i.e. which columns)
in your Marker Table represent the General Classes.
A general class is any class that is not a subtype of
another class.}

\item{multi.lvl}{set True if there are multiple subclasses of cells you
would like to classify.
Note: requires a manual entry table with columns specifying
levels and comparisons. Each column of this table should
represent one level of subclass within your General Classes.
Each value within a column should be two Class names
separated by a dash (-) showing the General-to-SubClass
relationship.
Example: To classify T-cells and then attempt to classify
those T-cells further into either CD8 or CD4 T-cells,
you would write a column named
"Level_1", then add "T_cell-CD8_T" and "T_cell-CD4_T" to
that column. Note that in this example, "T_cell"
is a General Class and "CD8_T" and "CD4_T" are not.}

\item{lvl.df}{Dataframe containing levels information as well as
parent-children designation (E.g. Tcells-CD4).
Required if Multi Level Classification is turned on.#'}

\item{reduction}{Choose among tsne, umap, and pca (Default: tsne)}

\item{nbins}{Number of bins for storing control features and analyzing
average expression (Default: 10)}

\item{gradient.ft.size}{Set size of axis labels on gradient
density plot of ModuleScore distribution
(Default: 6)}

\item{violin.ft.size}{Set size of axis labels on violin plot of
ModuleScore distribution (Default: 6)}

\item{step.size}{Set step size of distribution plots (Default: 0.1)}
}
\value{
List containing annotated dimension plot with ModuleScore
distribution of cell marker gene, Seurat Object with cell
classification metadata
}
\description{
Returns Seurat-class object with metadata containing
ModuleScores and Likely_CellType calls
}
\details{
Analyzed features are binned based on averaged expression;
control features are randomly selected from each bin.
}
\examples{
\dontrun{
modScore(
  object = seuratObject,
  marker.table = immuneCellMarkers,
  use.columns = c("CD4_T", "Treg", "Monocytes"),
  ms.threshold = c("CD4_T 0.1", "Treg 0.4", "Monocytes 0.3"),
  general.class = c("CD4_T", "Monocytes"),
  multi.lvl = FALSE
)

modScore(
  object = seuratObject,
  marker.table = immuneCellMarkers,
  use.columns = c("CD4_T", "Treg", "Monocytes"),
  ms.threshold = c("CD4_T 0.1", "Treg 0.4", "Monocytes 0.3"),
  general.class = c("CD4_T", "Monocytes"),
  multi.lvl = TRUE,
  lvl.df = parentChildTable
)
}
}
