% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Name_Clusters_by_Enriched_Cell_Type.R
\name{nameClusters}
\alias{nameClusters}
\title{Update metadata slot of Seurat-class object with custom labels and
provide plot with percentage of cell types}
\usage{
nameClusters(
  object,
  cluster.column,
  labels.column,
  cluster.identities.table,
  cluster.numbers,
  cluster.names,
  order.clusters.by = NULL,
  order.celltypes.by = NULL,
  interactive = FALSE
)
}
\arguments{
\item{object}{Seurat-class object with cluster IDs column and cell type
column present}

\item{cluster.column}{Column name containing cluster ID in the metadata slot
in the object}

\item{labels.column}{Column name containing labels (usually cell type) in the
metadata slot in the object}

\item{cluster.identities.table}{Data frame containing cluster IDs and custom
cluster labels}

\item{cluster.numbers}{Column name in cluster.identities.table
containing cluster numbers that match values in cluster.column}

\item{cluster.names}{Column name in cluster.identities.table
containing custom cluster labels}

\item{order.clusters.by}{Vector containing order of clusters in graph. Can
contain a subset of cluster numbers to plot that match at least some of
the values in the cluster.column. If NULL, use default order
(default is NULL)}

\item{order.celltypes.by}{Vector containing order of cell types in graph.
Can contain a subset of cell types to plot that match at least some of the
values in the labels.column. If NULL, use default order (default is NULL)}

\item{interactive}{If TRUE, draw plotly plot (default is FALSE)}
}
\value{
Returns Seurat-class object with updated meta.data slot containing
custom cluster annotation and a plot
}
\description{
Maps custom cluster names to Seurat Object cluster IDs and adds
cluster names to a new metadata column called \emph{Clusternames}. Provides a
dotplot of percentage of cell types within each cluster.
}
\examples{
\dontrun{
map_tbl <- data.frame(
  cluster_id = c("0", "1"),
  label = c("T cell", "B cell")
)
out <- nameClusters(
  object = anno_so,
  cluster.column = "seurat_clusters",
  labels.column = "celltype",
  cluster.identities.table = map_tbl,
  cluster.numbers = "cluster_id",
  cluster.names = "label"
)
}
}
