% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Name_Clusters_by_Enriched_Cell_Type.R
\name{NameClusters}
\alias{NameClusters}
\title{Updating cluster annotation in the meta.data slot of Seurat-class object}
\usage{
NameClusters(
  SO = seurat.object,
  metadata,
  cluster.identities.table,
  cluster.column.from.SO,
  cluster.names,
  cluster.numbers
)
}
\arguments{
\item{SO}{Seurat-class object with cluster a cluster IDs column and "Likely_CellTypes" column present}

\item{metadata}{a data.frame containing at least a cluster ID column; the column name will be used to retrieve clusters from the input Seurat Object, which then will be matched with the cluster IDs in the Cluster Identities Table}

\item{cluster.identities.table}{a data.frame with 2 columns - one with Cluster IDs (numeric) and the other with Cluster names}

\item{cluster.column.from.SO}{name of the cluster ID column present in the meta.data slot in the Seurat Object}

\item{cluster.names}{column containing cluster labels}

\item{cluster.numbers}{column containing cluster ID's (numeric)}
}
\value{
Returns Seurat-class object with updated meta.data slot containing custom cluster annotation or extracts a data.frame from the Seurat meta.data slot with already annotated cell types
}
\description{
Maps cluster IDs to cluster names and adds a new metadata column called \emph{Clusternames}. Seurat object is returned if all provided cluster IDs and cluster number from the Seurat Object meta.data match, otherwise a data.frame with counts of \emph{Likely_CellTypes} within each cluster is returned from the Seurat object.
}
