% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony.R
\docType{data}
\name{object}
\alias{object}
\title{Harmony Batch Correction from Singular Value Decomposed PCA}
\format{
An object of class \code{Seurat} with 3000 rows and 2000 columns.
}
\usage{
object
}
\arguments{
\item{seurat_object}{Seurat-class object}

\item{nvar}{Number of variable genes to subset the gene expression data by
(Default: 2000)}

\item{genes.to.add}{Add genes that might not be found among variably
expressed genes}

\item{group.by.var}{Which variable should be accounted for when running
batch correction}

\item{npc}{Number of principal components to use when running Harmony
(Default: 20)}
}
\value{
A list: adj.object with harmony-adjusted gene expression (SCT slot)
adj.tsne: harmonized tSNE plot
}
\description{
Adjusts cell embeddings and gene expression data to account for
variations due to user specified variable
}
\details{
Runs singular value decomposition on pearson residuals
(SCT scale.data) to obtain PCA embeddings. Performs harmony on
decomposed embedding and adjusts decomposed gene expression values
by harmonized embedding.
}
\examples{
\dontrun{
harmonyBatchCorrect(
  object = seurat,
  nvar = 2000,
  genes.to.add = c("Cd4", "Cd8a"),
  group.by.var = "Mouse_Origin",
  npc = 20
)
}
}
\keyword{datasets}
