% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Metadata.R
\name{plotMetadata}
\alias{plotMetadata}
\title{Plotting (i.e. coloring with) different columns of your Metadata Table}
\usage{
plotMetadata(
  object,
  samples.to.include,
  metadata.to.plot,
  columns.to.summarize,
  summarization.cut.off = 5,
  reduction.type = "tsne",
  use.cite.seq = FALSE,
  show.labels = FALSE,
  legend.text.size = 1,
  legend.position = "right",
  dot.size = 0.01
)
}
\arguments{
\item{object}{A combined Seurat Object with metadata to plot}

\item{samples.to.include}{Which samples you would like to include}

\item{metadata.to.plot}{The metadata columns from your Metadata table
you would like to plot}

\item{columns.to.summarize}{The columns you would like to summarize}

\item{summarization.cut.off}{Select the number of categories you want
to display, while marking all other cells as "other." Default is 5}

\item{reduction.type}{What kind of visualization you would like to use
to plot your cells and metadata (tsne, umap, pca). Default is tsne}

\item{use.cite.seq}{TRUE if you would like to plot Antibody clusters
from CITEseq instead of scRNA. Default is FALSE}

\item{show.labels}{Whether to add labels or not to your reduction map.
Default is FALSE}

\item{legend.text.size}{Customize the size of the legend text in your charts.
Default is 1}

\item{legend.position}{Select how you want to align your legend.
Default is "right"}

\item{dot.size}{The size of the dots displayed on the plot. Default os 0.01}
}
\value{
a data.frame extracted from the Seurat object and plot
}
\description{
For each column selected, this template will produce a plot
(UMAP/TSNE/PCA; your choice) using the data in that column to color the cells
}
\details{
This is a downstream template for the Single-cell RNA-seq workflow
(requires dataset where Filter/QC/SingleR annotations have been run first)
}
