% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_filter_QC_Plots.R
\name{Post_filter_QC}
\alias{Post_filter_QC}
\title{Post-filter QC Plots}
\usage{
<<<<<<< HEAD
Post_filter_QC(Seurat_Object, Image_type = "png")
=======
Post_filter_QC(Seurat_Object, Parallelize_Computation = F, Image_type = "png")
>>>>>>> 30b7146f0a89f5ab7f8ae790ea33038fe5ca58de
}
\arguments{
\item{Seurat_Object}{Please input a filtered Seurat Object. This should be the output from the Initial QC template.}

<<<<<<< HEAD
=======
\item{Parallelize_Computation}{Toggle true to parallelize your computations using Spark. Recommended for large datasets.}

>>>>>>> 30b7146f0a89f5ab7f8ae790ea33038fe5ca58de
\item{Image_type}{Remember that svgs are much larger than pngs, so we recommend doing everything first in png, then rerunning to output specific svgs as needed.}
}
\value{
Seurat Objects and QC plots that compares several QC parameters across samples after initial QC and filtering. This template is Step 2 in the basic Single-Cell RNA-seq workflow.
}
\description{
This set of plots compares several QC parameters across samples after initial QC and filtering. This template is Step 2 in the basic Single-Cell RNA-seq workflow.
}
\details{
Post-filter QC Plots template in single-cell-rna-seq-r4 NIDAP environment
from v47
}
