% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_filter_QC_Plots.R
\name{Post_filter_QC}
\alias{Post_filter_QC}
\title{Post-filter QC Plots}
\usage{
Post_filter_QC(object, imagetype = "png")
}
\arguments{
\item{object}{Please input a filtered Seurat Object. This should be the output from the Initial QC template.}

\item{imagetype}{Remember that svgs are much larger than pngs, so we recommend doing everything first in png, then rerunning to output specific svgs as needed.}
}
\description{
This set of plots compares several QC parameters across samples after initial QC and filtering.
}
\details{
This is Step 2 in the basic Single-Cell RNA-seq workflow. This template is used to evaluate different factors effecting data quality that should be taken into consideration for downstream processing.
}
