% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Names.R
\name{SampleNames}
\alias{SampleNames}
\title{Extracting cell identities from the meta.data slot of Seurat-class object}
\usage{
SampleNames(SO)
}
\arguments{
\item{SO}{Seurat-class object}
}
\value{
A table (data.frame) with the number of cells per each sample name (column names)
}
\description{
Returns tabulation of cell identities present in the orig_ident column
}
\details{
It is recommended to run this function after doing SingleR annotations and before doing any visualizations/downstream analysis that requires inputting sample names.
}
