% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_tSNE.R
\name{tSNE3D}
\alias{tSNE3D}
\title{Plot 3D TSNE from Seurat Object and saves as html file}
\usage{
tSNE3D(
  object,
  color.variable,
  label.variable,
<<<<<<< HEAD
  filename = "plot.html",
=======
  fileName = "plot.html",
>>>>>>> 30b7146f0a89f5ab7f8ae790ea33038fe5ca58de
  save.plot = FALSE,
  npcs = 15
)
}
\arguments{
\item{object}{Seurat-class object}

\item{color.variable}{Metadata column to use for color}

\item{label.variable}{Metadata column to use for label}

<<<<<<< HEAD
\item{filename}{Filename for saving plot (default is "plot.html")}
=======
\item{fileName}{Filename for saving plot (default is "plot.html")}
>>>>>>> 30b7146f0a89f5ab7f8ae790ea33038fe5ca58de

\item{save.plot}{Save plot (default is FALSE)}

\item{npcs}{Number of PC's (default is 15)}
}
\description{
Returns 3D tSNE plot for seurat object
}
\details{
This method provides visualization of 3D tSNE plot
}
