% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_Plots_by_Metadata.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin Plot by Metadata}
\usage{
violinPlot(
  object,
  assay,
  layer,
  genes,
  group,
  facet_by = "",
  filter_outliers = F,
  outlier_low = 0.05,
  outlier_high = 0.95,
  jitter_points,
  jitter_dot_size
)
}
\arguments{
\item{object}{Seurat-class object}

\item{assay}{Assay to extract gene expression data from (Default: SCT)}

\item{layer}{Slot to extract gene expression data from (Default: scale.data)}

\item{genes}{Genes to visualize on the violin plot}

\item{group}{Split violin plot based on metadata group}

\item{facet_by}{Split violin plot based on a second metadata group}

\item{filter_outliers}{Filter outliers from the data (TRUE/FALSE)}

\item{outlier_low}{Filter lower bound outliers (Default = 0.05)}

\item{outlier_high}{Filter upper bound outliers (Default = 0.95)}

\item{jitter_points}{Scatter points on the plot (TRUE/FALSE)}

\item{jitter_dot_size}{Set size of individual points}
}
\value{
violin ggplot2 object
}
\description{
Create violin plot of gene expression data across groups
}
\details{
Takes in a list of genes inputted by the user, displays violin plots
of genes across groups from a layer-assay with (optional) outliers
removed. Can also choose to scale or transform expression data.
}
\examples{
\dontrun{
violinPlot(
  object = seurat,
  assay = "SCT",
  layer = "data",
  genes = c("Cd4", "Cd8a"),
  group = "celltype",
  facet_by = "orig.ident",
  filter_outliers = TRUE,
  jitter_points = TRUE,
  jitter_dot_size = 0.5
)
}
}
