% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_Plots_by_Metadata.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{Violin Plot by Metadata}
\usage{
ViolinPlot(
  so,
  assay = "SCT",
  slot = "scale.data",
  ident_of_interest,
  groups_of_interest,
  genes_of_interest,
  filter_outliers = FALSE,
  scale_data = TRUE,
  log_scale_data = FALSE,
  reorder_ident = TRUE,
  rename_ident = "",
  ylimit = 0,
  plot_style = "grid",
  outlier_low_lim = 0.1,
  outlier_up_lim = 0.9,
  jitter_points = FALSE,
  jitter_width = 0.05,
  jitter_dot_size = 0.5,
  print_outliers = TRUE
)
}
\arguments{
\item{assay}{Name of the assay to extract gene expression data from}

\item{slot}{Name of the slot within the assay to extract gene expression data from}

\item{ident_of_interest}{Split violin plot based on a column from the seurat object metadata}

\item{groups_of_interest}{Include only a specific subset from ident_of_interest}

\item{genes_of_interest}{Genes to visualize on the violin plot}

\item{filter_outliers}{Filter outliers from the data (TRUE/FALSE)}

\item{scale_data}{Scale data from 0 to 1 (TRUE/FALSE)}

\item{log_scale_data}{Transform data onto a log10 scale (TRUE/FALSE)}

\item{reorder_ident}{Numeric data will be ordered naturally by default. Toggling this option will order the groups to match the group list if non-numeric, and will have no effect if otherwise.}

\item{rename_ident}{Give alternative names to ident_of_interest displayed on the graph}

\item{ylimit}{Y-axis limit}

\item{plot_style}{Choose between grid, labeled, and row}

\item{outlier_low_lim}{Filter lower bound outliers (default = bottom 10 percentile)}

\item{outlier_up_lim}{Filter upper bound outliers (default = bottom 90 percentile)}

\item{jitter_points}{Scatter points on the plot (TRUE/FALSE)}

\item{jitter_width}{Set spread of jittered points}

\item{jitter_dot_size}{Set size of individual points}

\item{print_outliers}{Print outliers as points in your graph that may be redundant to jitter}

\item{seurat_object}{Seurat-class object}
}
\value{
violin ggplot2 object
}
\description{
Returns a violin plot of gene expression data that is partitioned by user defined metadata column
}
\details{
Takes in a list of genes inputted by the user, displays gene expression information in particular slot-assay with (optional) outliers removed
}
