% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_Plots_by_Metadata.R
\name{violinPlot_mod}
\alias{violinPlot_mod}
\title{Violin Plot by Metadata}
\usage{
violinPlot_mod(
  object,
  assay,
  slot,
  genes,
  group,
  facet_by = "",
  filter_outliers = F,
  outlier_low = 0.05,
  outlier_high = 0.95,
  jitter_points,
  jitter_dot_size
)
}
\arguments{
\item{object}{Seurat-class object}

\item{assay}{Assay to extract gene expression data from (Default: SCT)}

\item{slot}{Slot to extract gene expression data from (Default: scale.data)}

\item{group.by}{Split violin plot based on metadata group}

\item{group.subset}{Include only a specific subset from group.by}

\item{genes.of.interest}{Genes to visualize on the violin plot}

\item{filter.outliers}{Filter outliers from the data (TRUE/FALSE)}

\item{scale.data}{Scale data from 0 to 1 (TRUE/FALSE)}

\item{log.scale.data}{Transform data onto a log10 scale (TRUE/FALSE)}

\item{reorder.ident}{Numeric data will be ordered naturally by default.
Toggling this option will order the groups to match the
group list if non-numeric, and will have no effect if
otherwise.}

\item{rename.ident}{Give alternative names to group.by displayed on
the graph}

\item{ylimit}{Y-axis limit}

\item{plot.style}{Choose between grid, labeled, and row}

\item{outlier.low.lim}{Filter lower bound outliers (Default = 0.1)}

\item{outlier.up.lim}{Filter upper bound outliers (Default = 0.9)}

\item{jitter.points}{Scatter points on the plot (TRUE/FALSE)}

\item{jitter.width}{Set spread of jittered points}

\item{jitter.dot.size}{Set size of individual points}

\item{print.outliers}{Print outliers as points in your graph that may be
redundant to jitter}
}
\value{
violin ggplot2 object
}
\description{
Create violin plot of gene expression data across groups
}
\details{
Takes in a list of genes inputted by the user, displays violin plots
of genes across groups from a slot-assay with (optional) outliers
removed. Can also choose to scale or transform expression data.
}
